fix(patina): allow dotted data table slots
