+++ crates/vize_patina/src/rules/vue/valid_v_slot.rs
@@ -252,7 +252,14 @@ fn has_component_parent(ctx: &LintContext) -> bool {
 fn static_slot_name(directive: &DirectiveNode) -> Option<String> {
     match &directive.arg {
         None => Some("default".into()),
-        Some(ExpressionNode::Simple(arg)) => Some(arg.content.to_compact_string()),
+        Some(ExpressionNode::Simple(arg)) => {
+            let mut name = arg.content.to_compact_string();
+            for modifier in &directive.modifiers {
+                name.push('.');
+                name.push_str(modifier.content.as_str());
+            }
+            Some(name)
+        }
         Some(ExpressionNode::Compound(_)) => None,
     }
 }
@@ -264,84 +271,4 @@ mod location_tests;
 mod mixed_default_tests;
 
 #[cfg(test)]
-mod tests {
-    use super::ValidVSlot;
-    use crate::linter::Linter;
-    use crate::rule::RuleRegistry;
-
-    fn create_linter() -> Linter {
-        let mut registry = RuleRegistry::new();
-        registry.register(Box::new(ValidVSlot));
-        Linter::with_registry(registry)
-    }
-
-    #[test]
-    fn test_valid_default_slot() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<MyComponent v-slot="{ item }">{{ item }}</MyComponent>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 0);
-    }
-
-    #[test]
-    fn test_valid_named_slot_template() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<MyComponent><template #header>Header</template></MyComponent>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 0);
-    }
-
-    #[test]
-    fn test_valid_default_slot_argument_on_component() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<MyComponent v-slot:default="{ item }">{{ item }}</MyComponent>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 0);
-    }
-
-    #[test]
-    fn test_valid_dynamic_component_slot() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<component :is="to ? NuxtLinkLocale : 'button'" #="scoped">{{ scoped }}</component>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 0);
-    }
-
-    #[test]
-    fn test_invalid_on_html_element() {
-        let linter = create_linter();
-        let result = linter.lint_template(r#"<div v-slot:header></div>"#, "test.vue");
-        assert_eq!(result.error_count, 1);
-    }
-
-    #[test]
-    fn test_invalid_named_slot_on_component() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<MyComponent v-slot:header>Header</MyComponent>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 1);
-    }
-
-    #[test]
-    fn test_valid_multiple_named_slots() {
-        let linter = create_linter();
-        let result = linter.lint_template(
-            r#"<MyComponent>
-                <template #header>Header</template>
-                <template #footer>Footer</template>
-            </MyComponent>"#,
-            "test.vue",
-        );
-        assert_eq!(result.error_count, 0);
-    }
-}
+mod tests;
